// components/spreadsheetInput/spreadsheetInputTargets.enum.ts
export enum SpreadsheetInputTargets {
  root = 'root',
  rowIndicator = 'rowIndicator',
  rowActionButtonWrapper = 'rowActionButtonWrapper',
  copyRowActionButton = 'copyRowActionButton',
  cutRowActionButton = 'cutRowActionButton',
  pasteRowActionButton = 'pasteRowActionButton',
  insetAboveRowActionButton = 'insetAboveRowActionButton',
  insertBelowRowActionButton = 'insertBelowRowActionButton',
  pasteAboveRowActionButton = 'pasteAboveRowActionButton',
  clearRowActionButton = 'clearRowActionButton',
  deleteRowActionButton = 'deleteRowActionButton',
}